requirements.txt
flask
langchain==0.3.27
langgraph==0.6.7
langchain-core==0.3.76
langchain-mistralai==0.2.11
python-dotenv
gunicorn
langgraph-checkpoint-sqlite
flask-cors